Cleanup, plus added type-wide subaction calls
Previously you had to define each subaction with a preceeding action name
that it was attributed to.  Now, if an action calls out for a subaction,
such as "object_updated", it will now check for $action_object_updated,
and if that does not exist, will check for it without a preceeding action,
so the sub object_updated would apply to any actions that make that call,
unless the specific action overrode this with a named subaction.